run deps outside of container

+++ app/scripts/run-dependencies.sh
@@ -13,21 +13,25 @@ docker run --rm \
     $registry/montagu-cert-tool:$cert_tool_version \
     gen-keypair /workspace
 
-# prepare to create orderly db
-rm $PWD/demo -rf
-rm $PWD/git -rf
-mkdir $PWD/demo
 ORDERLY_IMAGE="vimc/orderly:master"
 OW_MIGRATE_IMAGE="vimc/orderlyweb-migrate:master"
 OW_CLI_IMAGE="vimc/orderly-web-user-cli:master"
 
-#create orderly db
-docker pull $ORDERLY_IMAGE
-docker run --rm --entrypoint create_orderly_demo.sh -v "$PWD:/orderly" -u $UID -w /orderly $ORDERLY_IMAGE .
+if [[ -d demo ]]
+then
+  echo "Orderly demo folder already exists, not re-creating it."
+else
+  docker pull $ORDERLY_IMAGE
+  docker run --rm --entrypoint create_orderly_demo.sh -v "$PWD:/orderly" -u $UID -w /orderly $ORDERLY_IMAGE .
+fi
 
 # migrate to add orderlyweb tables
 docker pull $OW_MIGRATE_IMAGE
 docker run --rm -v "$PWD/demo:/orderly" $OW_MIGRATE_IMAGE
+
+# add users manage permission to test user for Orderly Web
+docker run -v $PWD/demo:/orderly $OW_CLI_IMAGE add-users test.user@example.com
+docker run -v $PWD/demo:/orderly $OW_CLI_IMAGE grant test.user@example.com */users.manage
 
 # Run the APIs and database
 docker-compose pull
@@ -43,10 +47,6 @@ docker exec montagu_db_1 montagu-wait.sh
 migrate_image=$registry/montagu-migrate:$MONTAGU_DB_VERSION
 docker pull $migrate_image
 docker run --network=montagu_default $migrate_image
-
-#Add users manage permission to test user for Orderly Web
-docker run -v $PWD/demo:/orderly $OW_CLI_IMAGE add-users test.user@example.com
-docker run -v $PWD/demo:/orderly $OW_CLI_IMAGE grant test.user@example.com */users.manage
 
 #start orderly web
 docker exec montagu_orderly_web_1 mkdir -p /etc/orderly/web
